tavern_server: auth, stage 1: implement
This commit contains most of the eventual authentication code, plus a
bunch of foundational code that made this and will make writing later
pieces easier.

Some notes:
- This only contains code for user registration, so far.
- This commit is missing tests and documentation.

As such, this commit is happening so I can make a pull request and get
the code in general looked at while I work on adding comments and tests.
The library builds; I have not bothered to check the executable yet.