tavern_server: create initial passing db tests
Two integration tests have been implemented: the script applies, and the
script applies twice, both without error.

Due to limitations in `sqlx`, namely that you can't run multiple
commands in a single query, I have opted to assume that the sql files
follows a particular format -- one empty line (including no spaces/tabs)
between two commands. It is fragile, but it works for now. I intend to
open an issue to allow multiple commands in a single query to make this
less fragile, but this works for now.

I also had to make a number of fixes to the sql file in order to get the
tests to pass:

- There were tables that depended on tables that were not yet defined
- Formatting was off and caused a few errors related to misplaced commas
- `CREATE TYPE IF NOT EXISTS` is not a valid syntax. What currently
  exists is the simplest way to achieve the same thing and is also the
  cause for the fragile script parsing (more on that later).

I also opted to make a fix to make the field names more consistent: the
`PRIMARY KEY` id field for each table is called `id` and all foreign
keys are `table_id`.

To make it easier to test the database, which requires either
environment variables to be defined or commandline arguments to be
provided, the added `Makefile` loads the file `.env`, which should
contain the necessary environment variables, like so:

```
ENVIRONMENT_VARIABLE=value
OTHER_VARIABLE="something else"
```