impl TryFromRow and TryFromUuid for Pathfinder types
Created two macros to derive the two traits on a given struct, because
the general process of getting a row from the database and making a
struct from it is mostly the same for all types, with some odd cases.
These are covered with a number of attributes that inform how the traits
are implemented.

In the future, some of those attributes should be able to be removed in
favor of parsing the field's type, but that's too much work for the
current deadline we have, and more attributes is easier to implement.

There is no guarantee the generated code even works: that's the upcoming
commit(s).

[TVRN-10]