migration to diesel pt 1
This commit adds most, if not all, of the diesel-related annotations for
types in tavern_pathfinder. Due to needing to `use` enum types derived
in tavern_pathfinder, the diesel migrations are stored in tavern_server.
So far, this also simpilifies the amount of custom macros needed in
tavern_derive.

Committing now as a good checkpoint to use moving forward, starting with
creating a patch file for `schema.rs` to add the custom enum types.