migration to diesel pt 2
Hundreds of compiler errors have been fixed (though many were the same
errors in multiple contexts). Currently all that remains is implementing
the tavern-related DB traits on objects, which the custom derive macros
will help implement them on the DB____ types, which helps with
implementing the actual Pathfinder types.

Among other changes, of note is the addition of a patch file for
schema.rs to add imports for the enums created by diesel-derive-enum.

The biggest change is that all of the code was merged under
tavern_server. While we wanted the Pathfinder types in a separate
library, using that with Diesel just added too much complexity.