Switch to Diesel (#23)
* impl TryFromRow and TryFromUuid for Pathfinder types

Created two macros to derive the two traits on a given struct, because
the general process of getting a row from the database and making a
struct from it is mostly the same for all types, with some odd cases.
These are covered with a number of attributes that inform how the traits
are implemented.

In the future, some of those attributes should be able to be removed in
favor of parsing the field's type, but that's too much work for the
current deadline we have, and more attributes is easier to implement.

There is no guarantee the generated code even works: that's the upcoming
commit(s).

[TVRN-10]

* migration to diesel pt 1

This commit adds most, if not all, of the diesel-related annotations for
types in tavern_pathfinder. Due to needing to `use` enum types derived
in tavern_pathfinder, the diesel migrations are stored in tavern_server.
So far, this also simpilifies the amount of custom macros needed in
tavern_derive.

Committing now as a good checkpoint to use moving forward, starting with
creating a patch file for `schema.rs` to add the custom enum types.

* migration to diesel pt 2

Hundreds of compiler errors have been fixed (though many were the same
errors in multiple contexts). Currently all that remains is implementing
the tavern-related DB traits on objects, which the custom derive macros
will help implement them on the DB____ types, which helps with
implementing the actual Pathfinder types.

Among other changes, of note is the addition of a patch file for
schema.rs to add imports for the enums created by diesel-derive-enum.

The biggest change is that all of the code was merged under
tavern_server. While we wanted the Pathfinder types in a separate
library, using that with Diesel just added too much complexity.

* run cargo fmt

* diesel migration: implement database traits

Not all operations are implemented on Effect and Feat, but the required
Get* methods are, automatically. `class.rs` and `character.rs` have not
yet been touched.

* (mostly) finish db impl for types

This commit finishes enough impls that the server compiles without
errors. Some types almost certainly need to have Insert/Update/Delete
implemented for them, but that can wait until they are needed.