add api traits, fix tests
All database tests now pass again after the change to Diesel.

Added API traits to make mapping the API more consistent. Some code was
refactored/changed to support that, and some changes were attempted to
be made but then had to be commented out because it was not quite
working and we have a class deadline to get *something* on the API
working.